examples/agent_evaluation_dspy/requirements.txt
eggai==0.1.16
dspy==2.5.43
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.2